requirements.txt
numpy
matplotlib
scipy.integrate